+++ CHANGELOG.md
@@ -9,5 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 ### Fixed
+
+- XP fluid is now the same fluid across every mod that makes it (FTBTesting/Testing-Issues#4204)
+- Ender IO's XP machines and tools store experience again (FTBTesting/Testing-Issues#4204)
+
 ### Removed
 ### Changed